pyproject.toml
[project]
name = "todo-fastapi-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi2postman>=0.1.2",
    "fastapi[standard]>=0.115.12",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
]
